+++ CHANGELOG.md
@@ -5,7 +5,19 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.24 - Unreleased
+## Version 0.2.25 - Unreleased
+
+### Added
+* Can now load do delayed loading of data in auxiliary space. (throws an error
+  if multiple auxiliary streams are mixed)
+* issubset issuperset for FusedChannelSpec
+* union for FusedChannelSpec and ChannelSpec
+* `find_asset_obj` for CocoImage
+* Added quantization spec. CocoImage.delay will now return a dequantized image
+  if quantization metadata is available.
+
+
+## Version 0.2.24 - Released 2022-03-15
 
 ### Added
 * verbose flag to perterb function